+++ index.js
@@ -9,7 +9,10 @@ const app = express()
 // middlewares
 app.use(cookieParser())
 app.use(express.json())
-app.use(express.urlencoded({ extended: true }))
+app.use(express.urlencoded({ extended: false }))
+
+// routes
+app.use('/api/users', require('./routes/users'))
 
 // serving static files
 app.use(express.static(__dirname + '/public'))