+++ models/User.js
@@ -2,9 +2,9 @@ const mongoose = require('mongoose')
 const Schema = mongoose.Schema
 
 const UserSchema = new Schema({
-  name: { type: String, required: true },
+  username: { type: String, required: true },
   email: { type: String, required: true, unique: true },
-  password: { type: String, required: true, min: 8, max: 30 },
+  password: { type: String, required: true, min: 6, max: 30 },
   date: { type: Date, default: Date.now },
 })